site_monitor/requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
kafka-python==2.0.2
psycopg2-binary==2.8.6
requests==2.25.1
urllib3==1.26.3
